Fixed bugs with brackets

+++ app/src/main/java/abcd/com/basiccalculator/Evaluators/PostFixCal.java
@@ -2,23 +2,24 @@ package abcd.com.basiccalculator.Evaluators;
 
 import java.util.Stack;
 
+
 public class PostFixCal {
     private String input;
     private String total;
-    float subtotal;
-    private Stack<Float> theStack;
+    double subtotal;
+    private Stack<Double> theStack;
     public PostFixCal(String in) {
         input = in;
         theStack = new Stack<>();
     }
 
     public String calculate() {
-        float ls;
-        float rs;
+        double ls;
+        double rs;
         String[] values = input.split("\\s+");
         for (String value : values){
             if (!value.matches("^[\\/\\+\\-\\*]$")){
-                theStack.push(Float.parseFloat(value));
+                theStack.push(Double.parseDouble(value));
             }
             else {
                 subtotal = 0;
@@ -49,7 +50,7 @@ public class PostFixCal {
             total = Integer.toString((int)subtotal);
         }
         else{
-            total = Float.toString(subtotal);
+            total = Double.toString(subtotal);
         }
         return total;
     }
